Fix CAN ID filtering, type truncation, and TransferData positive response
- ISOTP: Add CAN ID filter rejection in isotp_receive_frame() to prevent
  non-filtered CAN frames from corrupting the ISOTP state machine. Filter
  list expanded to include functional request ID (0x18DBFFF0).
- UDS diagnostic: Change len parameter from uint8_t to uint16_t to fix
  truncation of reassembled ISOTP data >255 bytes (TransferData).
- TransferData: Add positive response (0x76 + blockSequenceCounter) after
  successful data reception. Previously no response was sent.
- Flash download: Fix CRC loop counter from uint8_t to uint16_t to prevent
  infinite loop with data blocks >255 bytes.
- System: Add async UDS RX processing via ISR flag (g_uds_rx_pending).
- RoutineControl (0x31): Force positive response (0x71) for unknown RID.
- Clean encode conversion (UTF-16LE to ASCII) for header files.

+++ JC_FA2_v.17.32_20251203/USER/UDS/uds_diagnostic.h
@@ -142,7 +142,7 @@ typedef struct
 void uds_init(void);
 void uds_ms_update(void);
 void uds_process(void);
-int8_t uds_receive_handler(uint8_t channel, uint32_t can_id, uint8_t* data, uint8_t len);
+int8_t uds_receive_handler(uint8_t channel, uint32_t can_id, uint8_t* data, uint16_t len);
 int8_t uds_send_response(uint8_t channel, uint8_t sid, uint8_t* data, uint8_t len);
 int8_t uds_send_negative_response(uint8_t channel, uint8_t sid, uint8_t nrc);
 int8_t uds_send_response_pending(uint8_t channel, uint8_t sid);
